pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "line-bot-sdk>=3.14.5",
    "psycopg2-binary>=2.9.10",
    "tlgbotfwk>=0.4.61",
    "twilio>=9.4.5",
    "oauthlib>=3.2.2",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.0.1",
]
